refactor(events): tie outbox payload types to EventPayloads
Extract EventPayloads to event-payloads.ts so outbox-event-writer does not
import the package barrel, and model OutboxWriteEvent as a discriminated
union on eventName. createOutboxWriter no longer casts payload on insert.

Made-with: Cursor
